A new and improved front-end driver.  Custom instrumentation passes can be added by subclassing a basic gcc driver.
Output files are not quite correct yet when using "-c" or "-S" without
"-o".  We are getting a default final output file name based on the
name of the module being compiled, but in general this is a temporary
file with a randomly generated name.  Somehow I need to propagate
forward enough information about the original source file's name so
that an explicit output name can be generated accordingly.