Site descriptions have gotten too big to embed as literal C strings. GCC crashes when string literals get up around 10MB or so.  :-P
The new strategy is to spit out the site description text into a
separate file, and then use "objcopy --add-section" to copy into a new
section within the object file.  Doing this requires that we hook in
on the assembler, since the assembler is the one who gets to see the
object file.

Control flow graph embedding has been converted to use the same
approach.