Add support for balancing out all paths in an acyclic region, per ideas discussed with Cormac Flanagan at PLDI 2004.
The key to making this work in a sane way is to find places we can put
the extra dummy sites which only affect an immediately preceding "if"
or "switch" statement and cannot be reached in any other way.  Though
I've tested this quite a bit, the handling of "switch" statements is
quite subtle and may still harbor bugs.  Any such bugs should be
detected, though, during a second path weighing pass where we assert
that everything is well balanced.