Hairy reworking of path balancer, along with a new regression test representing the bug that prompted it.
We cannot prepatch "if" statements and try to remember what the "then"
and "else" clauses were, because later patching of a "switch"
statement might want to modify parts of those clauses to redirect
"goto" statements.  So instead we just collect a simple list of
statements and do a second pattern matching scan across this list
when it's time to balance paths.

In an effort to keep sane, move some of the "switch" case rewriting
into the prepatch phase.  Specifically, rewrite all cases so that they
have explicit "goto" instead of implicit fall-through onto the next
case.  Doing this earlier ensures that any "goto" statements we
introduce will be picked up and patched properly as jumps.

Unfortunately this means that we need to introduce explicit "default"
cases even earlier than that, because adding "default" cases
introduces "break" statements and the loop removal phase happens
fairly early.  So introduction of explicit "default" cases needs to
happen early too.

Ugh.