When balancing paths coming out of an "if" statement, we need to know that both the "then" and "else" blocks are non-empty so that we can ask the first statement in each block for its weight and so that we know we won't have anything falling through without hitting either clause.  So we used to insert an empty statement into each block if that block was empty.
However, that's not quite right.  If one of these blocks was
previously nonempty, but the first statement contained an acyclic
region header, then asking it for its weight would give us the weight
of that header's acylic region.  What we really want in that case is
the weight from the perspective of the "if" statement, which here
would be zero.

So now we just unconditionally add an empty statement to the start of
each clause.  This guarantees that we have a successor statement which
is in the same acyclic region as the "if" statement.