Switch to using XSLT to convert DocBook into XHTML.  It produces more standards-compliant output, it's faster than DSSSL, and I understand how it works well enough to make small tweaks when needed.
Add a CSS stylesheet to be applied to the converted XHTML.  It's tiny
for now, but will probably grow.  I may just want to grab a nice
stylesheet off the web somewhere.

Add a detailed discussion of the static site information file format.
More file format documentation is on the way....