Generate HTML right here, instead of in a subdirectory.
Avoid using a stamp file for HTML conversion.  Near as I can tell we
can depend on xsltproc to write "index.html" last, and only if it made
it all the way through.  And just in case that fails, we've told
"make" to delete files it was generating in the event of an error.