Avoid embedding the full cc1 path in the generated RPM spec file. That makes the spec file platform specific, which in turn makes the SRPM platform specific.
Instead use the %(shell command) expansion facility to ask gcc for its
cc1 path.  This is evaluated when a binary RPM is built, which means
that the SRPM remains platform neutral.

This trick works fine on Fedora Core 2 (rpm-4.3.1).  Haven't yet tried
it on Fedora Core 1 (rpm-4.2.1), but that's the next thing to do.