Improve clustering of nodes within subgraphs.  This should especially make it easier to recognize individual functions.
Link each caller node to the entry node of the callee.  Linking to the
entire callee subgraph is not supported.  I tried using "ltail" to
link to the entry node of the callee but clip the line at the edge of
the callee subgraph, but that yields confusing error messages from
dotty.