Change position of sampler support libraries on linker command line.
When "--threads" is active on Fedora Core 4, putting sampler librares
with the regular libraries yields executables which crash during
startup, and even crash when given to "ldd".  There seems to be some
issue with libgcc_s, but I'm not entirely sure what.  Moving the
libraries from "*lib" to "*libgcc" puts my libraries before libgcc_s,
which in turn seems to fix the problem.