Use installed Python libegg tray icon wrapper if available.  Fedora Core 4 finally provides a standard gnome-python2-libegg package with this functionality.
Otherwise, fall back on our own copy of this code as before.

Detecting the desired Python module is tricky.  The obvious test,
("import egg.trayicon") aborts if $DISPLAY is unset.  This happens,
for example, when building RPMs.  So instead we need to use a more
primitive interface to Python's module loading system to find the
egg.trayicon module without actually loading it.