Hook up both uninitialized variable analyses at the same time. Compare their results to see if they behave identically.  I've tested this on bc, exif, and ccrypt and we do indeed get the same results for all program sources.  Cool!  I am now much more confident that both implementations are correct.
Run each of the two analyses *twice*, actually, and time it each time.
I run them in a sequence like <A, B, B, A> to avoid penalizing either
implementation for having to go first with a cold cache.  Comparing
the first B run against the second A run should be more or less fair.
Compared in this manner, Initialized is about 3 times faster than
MustBeUninitialized.  Swapping the order to <B, A, A, B> gives
Initialized a factor of 3.3 speedup over MustBeUninitialized.

This is a slightly dodgy comparison, because MustBeUninitialized uses
hash table lookups for post-analysis later queries while Initialized
uses a binary search tree.  But the difference in analysis time,
coupled with the fact that the results are otherwise identical, has me
pretty convinced that ditching use-def analysis in favor of a simpler
approach is a net win.

I'll probably let this code settle in for a day or so.  If neither I
nor Mulhern can think of any problems, I'll eventually merge just the
faster Initialized module into the trunk.