Committing the first pass on my implication logger.
The basic plan is:

- keep a list of site id, constant pairs by scalarPairFinder.ml.
In place updating of the list is required because of the scalarPairFinder
visitor structure. The constants are stripped down to their underlying int64
representation.

- needed to update the counters.manager interface to return the site id. This
necessitated changing a number of implementations that use the addSite method
to explicitly throw away the site id.

- implications.ml contains the analysis. It is completely self-contained,
making no references to Cil types, much less sampler types. This will probably
change; most likely it will be wrapped in something larger that contains
more information. But lets see if it works in the context of single
compilation unit and a single variable before we get all carried away.