Minor bug fix: I have to dig to get the actual numeric representation of the constant so that I can compare it. I wasn't digging deep enough, i.e., I was ignoring the possibility that the expression was a cast of some kind. I now deal with that possibility. I explicitly assert false if it is anything other than a cast.
There's an argument for putting some or all of this in implications.ml
or counters.ml and yanking it out of scalarPairFinder.ml entirely.
In that case one would probably revert to the previous interface for
addSite and revert all the files that had to be updated to handle the new
interface.

It should be noted that the results of all this effort don't amount to much
without the --compare-constants flag.