Do the analysis in chunks. Since our simple approach can only relate the predicates at a single inspiration via implication we do the analysis and dump out the implication relations on a per-inspiration basis. This means that the size of the list we handle is O(n^2) where n is the number of instrumentation sites for any given inspiration.
This allows us to generate an implication file for our challenge file "bc.c".
This in spite of the fact that it looks like there may be as many as 1000
instrumentation sites for a given inspiration in some places.

It will be necessary to break the output into multiple files, otherwise reading
it will result in other unpleasant scalability issues.

cc1 runs out of memory while trying to compile the instrumented source.
which is only around 97,000 lines.

Maybe all files automatically generated by lexers (such as bc.c) should be
exempted from constant comparisons.