Collapse two "match" blocks into a single one that looks more deeply into the data structures being considered.
Use "0L" instead of "Int64.zero" for improved readability.

When checking to see if unsigned values are larger than 2^32, only to
that test once for each value.  Save the results of these comparisons
as booleans, and then pattern match on the booleans to decide how to
complete the comparison.  I like this style a bit better because it
makes it clear what the four cases are.