Don't build implications unless --sampler-implications flag is set.
If flag is set, implications are written to <filename>.impls.

Use of lazy required so that we dereference Implications.saveImplications in
a context where it has been set if it will be set, i.e., after the command
line arguments have been parsed.

Retooled implications calculations a bit to make interface with
scalarPairsFinder simpler. Gets same results as before.