Changes for using scons build directories.  Works with duplication=True, but not with duplication=False.  Not sure if it works with no build directory at all.
In any case, I've decided the hacks required to make this all work are
ugly enough that I'm not going to push it further for now.  I'm
checking this in to preserve the tweaks up to this point, in case I
decide to give it a second look at a later date.