Slightly experimental change from Bonobo to D-Bus for managing the activation and lifetime of the sampler tray monitor icon.
The Bonobo-based approach is giving me two icons under Fedora 8, in spite of the fact that this is supposed to be single-instance.  The whole Bonobo infrastructure is over-engineered for this anyway, and the general trend among GNOME apps is moving away from Bonobo and toward D-Bus anyway.

The D-Bus approach is working properly under Fedora 8: one icon, no more.  Also, D-Bus services get a nice notification signal when other D-Bus clients go away, which means that we can be more robuse about pulling the icon down even if a client's wrapper script crashes abnormally.  That's something I always wanted from Bonobo, but which was fiendishly difficult to achieve.

Haven't tried this under older-but-still-supported platforms (Fedora 6 & 7) yet.  That will come next.