Restore "/afs/cs.wisc.edu/p/cbi/public/build/cbiexp" as the default directory for the root of the CBI analysis tools tree.  However, this can be overridden by creating a "sampler/mining/analysis/local.mk" file like this:
     root = /u/f/l/fletchal/CBI/cbiexp/trunk

If that "local.mk" file is missing, we quietly fall back on
"/afs/cs.wisc.edu/p/cbi/public/build/cbiexp" as the default.

Subversion has been told to ignore "local.mk", so this is a good place
to put customizations that are not appropriate as permanent additions
to the source tree.