More cs.wisc.edu build tweaks:
On department machines, run pychecker using /usr/bin/python as the
interpreter, so that we find standard Python packages installed on the
system.  Otherwise it uses /s/python2/bin/python, which causes it to
miss the installed GNOME Python modules.

Also on department machines, skip the 'launcher' subdirectory
entirely.  The department's machines have a pre-0.80 version of the
Python D-Bus bindings, and we don't support that any more.  This is
particularly a problem for "launcher/tray", but I'm skipping the whole
launcher subsystem since we won't really be using it on department
machines anyway.