Rewriting pass that ensures that each statement accesses no more than one shared, mutable memory location.  Any statement which accesses more than one such location is rewritten as multiple simpler statements.  Each temporary variable added when doing this has its description set to the original source expression being held in that tempvar.  This should help us produce predicate descriptions that are more useful than simply using the generated tempvar names.
I've tested this fairly extensively in isolation, and it's working
well.  The only quirk is that it rewrites expressions more
aggressively than strictly necessary in "if" and "switch" conditions,
but really that's not going to be a problem at all and should not even
change the machine code that is ultimately generated.

I have *not* yet incorporated this code into the main instrumentor.
That will come next, and should be pretty easy.