Better static description information for atoms instrumentation sites. We now give two additional site-specific fields for these sites:
	1. the source expression representing the shared access
	2. "read" or "write" depending on the kind of access

The source expression always uses original program expressions,
avoiding the names of temporary variables.  The actual access is the
outermost (last) one in the given expression.  For example, suppose
"vector" and "slot" are both global variables.  Then "vector[slot]"
refers to reading the global array *after* the index has been
computed.  For the shared access required to compute the index in the
first place, the expression would just be given as "slot".

Guoliang, I am done with my instrumentor updates for now.  This code
is ready for you to use.