Taking the address of a variable does not access the value of that variable.  For example, if "x" is a global variable then "x" access shared memory but "&x" does not.
Likewise, taking the address of a pointer dereference does not access
the pointed-to memory.  For example, if "p" is a local pointer
variable, then "*p" would access the memory it points to, but "&(*p)"
does not.

Generalizing the pointer case, if "p" is a local pointer to a
linked-list node, then "p->next" would access the memory it points to,
but "&p->next" does not.  Rather, this is merely type-safe pointer
arithmetic: adding the offset of the ".next" field to the base value
in p.

This ought to fix Guliang's latest bug, which concerned code of the
form "active->ref = &node->next;".  In that case, "active->ref" is
dereferencing a pointer and is therefore accessing potentially-shared
memory.  But the left side, "&node->next", does not access shared
memory at all.