Convert most GUI elements to use GtkBuilder instead of libglade, as the later is deprecated in favor of the former.  Move from GnomeAbout to GtkAboutDialog at the same time, for the same reason.
Unfortunately the tray icon stays with libglade for now.  The
GtkBuilder documentation seems to suggest that it doesn't handle popup
menus in the version that shipped with Fedora 10, which we still
support.

Rename the file that holds documentation configuration information
from "config.xml" to "config.dtd".  It really is a bit of DTD code,
not an XML document.