Convert tray icon to use GtkBuilder as well, with the popup menu being embedded as a GtkUIManager.  This seems to work fine under Fedora 11, but I haven't actually tried it under Fedora 10 yet.
The only annoyance with this conversion is that the
gtk.Builder.connect_signals call really wants to find all signal
handlers in a single dictionary or instance.  So PopupMenu needs to
have a few extra methods which explicitly delegate down to
AboutDialog.  Under libglade, each of these classes could bind signal
handlers for its own widget subtree, but I see no way to do that with
GtkBuilder.