## Lines starting with '## ' will be removed from the log message. ## File(s) to commit recursively: ## . Allow explicitly selecting 32- or 64-bit counters in predicate counter tuples.  This is a build-time option; use "tuple_counter_bits=32" or "tuple_counter_bits=64" on the scons command line.  Default is "tuple_counter_bits=natural", which uses the platform's natural integer size.
It's critical that the run-time libraries and all instrumented code
agree on what size is being used.  sampler-cc will default to using
the appropriate size based on how the libraries were built.

If you are using the libraries without sampler-cc (i.e., if you are
Tristan), then you must define a preprocessor macro to select the
appropriate bit width before including any headers for the run-time
library.  Set CBI_TUPLE_COUNTER_BITS to "32", "64", or "natural".  If
not set, the default is "natural".