Add new flags to include/exclude instrumentation by source file and line number:
     --include-location=foo.c:20
     --exclude-location=bar.c:45

Filenames can be wildcards ("*"), though that seems unlikely to
actually be useful.  Line numbers cannot be wildcards; if you want to
match all lines in a given file, use the "--include-file=" and
"--exclude-file=" flags, which are still available as before.  

The location- and file-based flags all use the same
inclusion/exclusion patterns list.  First match wins, with default to
include, as before.

Alex, I've tested these *somewhat*, at least to confirm that flag
parsing is correct.  Please be alert to possible regressions or
malfunctions of the new logic!