Under D-Bus 1.4 (Fedora 14 beta), we get a NameOwnerChanged message immediately after starting up, apparently caused by something we do ourselves.  This tricks us into exiting immediately, to the dismay of any client that intended to communicate with us.
The fix is to not exit the server until at least one real client has left the bus.

Currently committing with lots of extra debug output in place.  I want to double-check this all under Fedora 13 to ensure it still works correctly there.

Small bonus fix #1: use "self.__clients", not "self.clients".  The former is what we claim to use in the __slots__ list, and nobody outside this class needs to see that client list anyway.

Small bonus fix #2: when registering for the NameOwnerChanged signal, use argument matching to restrict us to only seeing client departures.  Those are the only ones we want, so there's no sense being notified of new arrivals or renamings.