Revamp of our pthread_create hijacking strategy.
Using the "--wrap" linker flag turns out to be a bit brittle.  gcc-4.6.0 has its own __wrap_pthread_create implementation in libgcc.a, used as part of its runtime support for the "-fsplit-stacks" feature.  Unfortunately, "-lgcc" generally appears earlier in the link order than our libraries, which means that the redirected pthread_create ends up being captured by libgcc's __wrap_pthread_create instead of our own.

The new strategy is based on one Tristan is already using in his git tree copy of the sampler runtime libraries.  Instead of using "--wrap", we directly implement a function named pthread_create.  Within that function, we call dlsym with the special RTLD_NEXT handle to find the *next* definition of pthread_create in the search order after the current library.  That will typically be the one in libpthread.  For this to work, we just need to ensure that our pthread_create appears earlier in the link order than the one in libpthread.

I've applied the same change to our hijacking of fork.  There was no regression in how the old approach worked for fork, but I'd rather have us using just one function-hijacking strategy rather than two.

With this change, almost all of sampler-cc's regression tests pass under Fedora 15 beta.  The only remaining failure seems to be a pychecker issue.  The instrumentor itself seems to be working fine, though.