More Perl-to-Python porting.
Change substitution templates to use $var instead of @var@ so that we
can use the standard Python string.Template class.  Previously we had
a subclass of this that worked with @var@, but that no longer works
for some reason under Fedora 15.

This could all use more testing than it has at the moment, so be wary
of regressions!