Waiting for a notification service to appear on the session bus does not work as hoped.  It seems GNOME Shell is claiming the org.freedesktop.Notifications service but does not have the proper interface ready yet.
So give up on a bus-based solution to this.  Instead, if posting the note fails, set a 1-second timeout to try again.  Keep retrying once per second until the note is posted without throwing an exception.

This appears to fix a deterministically-repeatable failure under Fedora 17 caused when Nautilus is managing the desktop and GNOME Shell (not GNOME Panel) is running the session.

Will release this as sampler-1.9.1-1.