Treat "-l*" flags as though they were input files, so that they retain their original order relative to other input files.  This means that if you link with "foo.o -lz bar.o", then library "z" is searched after "foo.o" but before "bar.o".
This is consistent with GCC's behavior.  In fact, the example I gave
above is taken directly from the GCC documentation.

Fortunately this was actually pretty easy given the support already
(recently) added for tracking the source language for each input file.
We just need to treat "-l*" flags as though they were naming input
files that are already compiled down to the "object" language.