While recognizing flags, also record which compilation stages they should be used in.  Previously we were filtering out inapplicable flags much later, which felt almost like redundant parsing.
Still some ugliness in filtering out "-pthread" as a linker flag when
either "-nostartfiles" or "-nostdlib" was given.  Not sure what I want
to do to make this prettier.

Add support for "-save-temps=cwd" and "-save-temps=obj".  These gcc
flags are not actually supported by clang, though.  If clang starts
supporting them in the future, we will already be ready.