Remove ability to disable notification tray icon.
GNOME 3.8 has a generic ability to enable/disable per-application
notifications.  No sense in duplicating that work.

Unfortunately this does remove functionality for older GNOME
environments.  The alternative would be to make this conditional, or to
delay removing it entirely until GNOME 3.8 is the oldest supported
version.  But I'd rather simplify whenever possible, and it's not clear
that many (any?) users still require this feature.