CIL's Rmtmps.removeUnusedTemps code has gotten more aggressive about rewriting (mutating) statements in place.  This happened sometime between release 1.4.0 and 1.7.3.
Most notably for our purposes, this affects the threshold checks
inserted by backward-jump patching.  Previously we created one
statement implementing this check but inserted it into the AST in two
places: one in the fast clone and one in the slow clone.  When
removeUnusedTemps transforms either of these it therefore affects the
other.  Unfortunately that transformation is not actually appropriate
for both.  This led to a failure of the g-object-unref regression
test, which has several such threshold checks after various function
calls.

The fix is to avoid having two aliased references to the same
statement in the tree.  While that worked fine for 11+ years, it was
always a somewhat sketchy thing to do.  Instead, we insert the "real"
threshold check just once.  For the other clone, we simply insert a
freshly-created goto to jump to the real threshold check.  I'm
trusting the downstream native C compiler to untangle the rather
convoluted chains of branches and gotos that result.

This fixes all known regressions when compiling sampler-cc against CIL
1.7.3.  All regression tests now pass.