Create a new notification instance each time "enabled" toggles.
Reusing an existing Notification crashes under GNOME 3.10 as shipped
with Fedora 20 beta.  So apparently we have to create a new notification
each time we toggle the state.  This doesn't look quite as good as
recycling, which did work under GNOME 3.8 as shipped with Fedora 19.
But recycling with a crash is no good at all, so we do what we must.

Tested under GNOME 3.10 (Fedora 20 beta), GNOME 3.8 (Fedora 19), and
GNOME 3.6 (Fedora 18).