Add a pragma to suppress warnings about a warning-suppression pragma
When using GCC 8.1.1 on Fedora 28, we need to temporarily suppress
"cast-function-type" warnings.  However, GCC 4.8.2 on Ubuntu 14.04
doesn't know about this warning category.  It warns about the unknown
warning category, which we then treat as an error.  So now we need to
add another pragma to tell GCC to ignore warnings about unrecognized
pragmas.  This is goofy.