A few extra linker tweaks for Ubuntu 18.04
Some tests were failing because the instrumented test executable links
with some libraries "X" which in turn depend on other shared libraries
"Y".  When launching the executable, it was failing to find the "Y"
libraries even though "Y" and "X" live in the same directory, and that
directory was part of the executable's RPATH.  Setting RPATH to
"$ORIGIN" in the "X" libraries seems to fix this.  This wasn't needed
on Fedora 28, but apparently Ubuntu 18.04 has a different policy for
how RPATH applies to transitive shared-library dependencies.

When linking with no instrumentation at all, test executables were
also not linking with the "libsampler.so" shared library, because they
didn't really seem to need it.  However, we do still want to have that
shared library loaded at run time so that it can create an empty
reports file.  Adding "--no-as-needed" at the right spot in the
instrumenting linker's command line ensures that "libsampler.so" will
be linked into the instrumented executable.  "--no-as-needed" is
actually the default, but apparently Ubuntu 18.04 sneaks a
"--as-needed" in somewhere.  Fedora 28 does not.